// sourcecode/BTrak.Packages/snova-admin-module/admin-module-components/admin-components/src/lib/admin-module/models/hr-models/role-model.ts
export class RolesModel {
    roleId: string;
    roleName: string;
    list1:string;
    fullName:string;
  }

  export declare class RoleModel {
    roleId: string;
    roleName: string;
    features: string[];
    featureIds: string[];
    isArchived: boolean;
    featureIdXml: string;
    data: string;
    timeStamp: any;
    isDeveloper: boolean;
}